+++ amber/cli.py
@@ -4,7 +4,7 @@ import os
 from yaml import safe_load  # type: ignore
 from amber.automation.detic_image_labeler import DeticImageLabeler
 from amber.automation.nerf_3d_reconstruction import Nerf3DReconstruction
-from amber.dataset.rosbag2_dataset import Rosbag2Dataset
+from amber.dataset.images_dataset import ImagesDataset
 from amber.importer.video import VideoImporter
 from typing import Any, Callable
 
@@ -60,7 +60,7 @@ def run_detic_image_labaler_automation(args: Any) -> None:
     with open(args.task_description_yaml_path, "rb") as file:
         task_description = safe_load(file)
     labeler = DeticImageLabeler(args.task_description_yaml_path)
-    dataset = Rosbag2Dataset(args.rosbag_path, args.dataset_description_yaml_path)
+    dataset = ImagesDataset(args.rosbag_path, args.dataset_description_yaml_path)
     annotations = labeler.inference(dataset)
     labeler.write(dataset, "/detic_image_labeler", annotations, args.output_rosbag_path)
 
@@ -71,7 +71,7 @@ def run_nerf_3d_reconstruction_automation(args: Any) -> None:
     with open(args.task_description_yaml_path, "rb") as file:
         task_description = safe_load(file)
     reconstruction = Nerf3DReconstruction(args.task_description_yaml_path)
-    dataset = Rosbag2Dataset(args.rosbag_path, args.dataset_description_yaml_path)
+    dataset = ImagesDataset(args.rosbag_path, args.dataset_description_yaml_path)
     reconstruction.inference(dataset)
