+++ tests/read_image_ford.yaml
@@ -1,3 +1,2 @@
-dataset_type: image_only
 image_topics:
   - topic_name: /image_front_left
