+++ tests/read_image_vrx.yaml
@@ -1,4 +1,3 @@
-dataset_type: image_only
 image_topics:
   - topic_name: /wamv/sensors/cameras/front_left_camera_sensor/image_raw
     compressed: true
