+++ tests/read_images_soccer_goal.yaml
@@ -1,3 +1,2 @@
-dataset_type: image_only
 image_topics:
   - topic_name: /camera/image_raw
